+++ src/api/xtream_player_api.rs
@@ -18,7 +18,6 @@ use crate::repository::xtream_repository::{COL_CAT_LIVE, COL_CAT_SERIES, COL_CAT
                                            xtream_get_all, xtream_get_stored_stream_info, xtream_persist_stream_info};
 use crate::utils::{get_client_request};
 
-
 fn get_xtream_player_api_action_url(input: &ConfigInput, action: &str) -> Option<String> {
     match input.input_type {
         InputType::M3u => None,
@@ -33,7 +32,6 @@ fn get_xtream_player_api_action_url(input: &ConfigInput, action: &str) -> Option
 }
 
 fn get_xtream_player_api_info_url(input: &ConfigInput, cluster: &XtreamCluster, stream_id: i32) -> Option<String> {
-
     let (action, stream_id_field) = match cluster {
         XtreamCluster::Live => ("get_live_info", "live_id"),
         XtreamCluster::Video => ("get_vod_info", "vod_id"),
@@ -186,28 +184,6 @@ async fn xtream_player_api_timeshift_stream(
     xtream_player_api_stream(&req, &api_req, &_app_state, "timeshift", &username, &password, &action_path).await
 }
 
-async fn xtream_get_stream_info_response(app_state: &AppState, user: &UserCredentials,
-                                         target_name: &str, stream_id: &str,
-                                         cluster: &XtreamCluster) -> HttpResponse {
-    match FromStr::from_str(stream_id) {
-        Ok(xtream_stream_id) => {
-            if user.proxy == ProxyType::Redirect {
-                if let Some(target_input) = app_state.config.get_input_for_target(target_name, &InputType::Xtream) {
-                    if let Some(info_url) = get_xtream_player_api_info_url(target_input, cluster, xtream_stream_id) {
-                        return HttpResponse::Found().insert_header(("Location", info_url)).finish();
-                    }
-                }
-            }
-
-            match xtream_get_stream_info(app_state, target_name, xtream_stream_id, cluster).await {
-                Ok(content) => HttpResponse::Ok().content_type(mime::APPLICATION_JSON).body(content),
-                Err(_) => HttpResponse::Ok().content_type(mime::APPLICATION_JSON).body("{info:[]}"),
-            }
-        }
-        Err(_) => HttpResponse::BadRequest().finish()
-    }
-}
-
 async fn xtream_get_stream_info(app_state: &AppState, target_name: &str, stream_id: i32,
                                 cluster: &XtreamCluster) -> Result<String, Error> {
     if let Some(target_input) = app_state.config.get_input_for_target(target_name, &InputType::Xtream) {
@@ -237,6 +213,27 @@ async fn xtream_get_stream_info(app_state: &AppState, target_name: &str, stream_
     Err(Error::new(std::io::ErrorKind::Other, format!("Cant find stream with id: {}/{}/{}", target_name, &cluster, stream_id)))
 }
 
+async fn xtream_get_stream_info_response(app_state: &AppState, user: &UserCredentials,
+                                         target_name: &str, stream_id: &str,
+                                         cluster: &XtreamCluster) -> HttpResponse {
+    match FromStr::from_str(stream_id) {
+        Ok(xtream_stream_id) => {
+            if user.proxy == ProxyType::Redirect {
+                if let Some(target_input) = app_state.config.get_input_for_target(target_name, &InputType::Xtream) {
+                    if let Some(info_url) = get_xtream_player_api_info_url(target_input, cluster, xtream_stream_id) {
+                        return HttpResponse::Found().insert_header(("Location", info_url)).finish();
+                    }
+                }
+            }
+
+            match xtream_get_stream_info(app_state, target_name, xtream_stream_id, cluster).await {
+                Ok(content) => HttpResponse::Ok().content_type(mime::APPLICATION_JSON).body(content),
+                Err(_) => HttpResponse::Ok().content_type(mime::APPLICATION_JSON).body("{info:[]}"),
+            }
+        }
+        Err(_) => HttpResponse::BadRequest().finish()
+    }
+}
 
 async fn xtream_get_short_epg(app_state: &AppState, user: &UserCredentials, target_name: &str, stream_id: &str, limit: &str) -> HttpResponse {
     if let Some(target_input) = app_state.config.get_input_for_target(target_name, &InputType::Xtream) {
@@ -259,7 +256,7 @@ async fn xtream_get_short_epg(app_state: &AppState, user: &UserCredentials, targ
                             }
                             Err(err) => {
                                 error!("Failed to download epg {}", err.to_string());
-                                return HttpResponse::NoContent().finish()
+                                return HttpResponse::NoContent().finish();
                             }
                         }
                     }
@@ -298,7 +295,8 @@ async fn xtream_player_api(
                     }
                     "get_short_epg" => {
                         xtream_get_short_epg(_app_state, &user, target_name,
-                                             api_req.stream_id.trim(), api_req.limit.trim()).await
+                                             api_req.stream_id.trim(),
+                                             api_req.limit.trim()).await
                     }
                     _ => {
                         match match action {
@@ -310,8 +308,7 @@ async fn xtream_player_api(
                             "get_series" => xtream_get_all(&_app_state.config, target_name, COL_SERIES),
                             _ => Err(Error::new(std::io::ErrorKind::Unsupported, format!("Cant find action: {}/{}", target_name, action))),
                         } {
-                            Ok(maybe_file_path_or_content) => {
-                                let (path, content) = maybe_file_path_or_content;
+                            Ok((path, content)) => {
                                 if let Some(file_path) = path {
                                     serve_file(&file_path, req).await
                                 } else if let Some(payload) = content {